some adjusts
