+++ components/circle/sources/new-comment-posted/new-comment-posted.mjs
@@ -1,8 +1,8 @@
-import { DEFAULT_POLLING_SOURCE_TIMER_INTERVAL } from "@pipedream/platform";
-import circle from "../../circle.app.mjs";
+import common from "../common/base.mjs";
 import sampleEmit from "./test-event.mjs";
 
 export default {
+  ...common,
   key: "circle-new-comment-posted",
   name: "New Comment Posted",
   description: "Emit new event each time a new comment is posted in the selected community space.",
@@ -10,33 +10,10 @@ export default {
   type: "source",
   dedupe: "unique",
   props: {
-    circle,
-    db: "$.service.db",
-    timer: {
-      type: "$.interface.timer",
-      default: {
-        intervalSeconds: DEFAULT_POLLING_SOURCE_TIMER_INTERVAL,
-      },
-    },
-    communityId: {
-      propDefinition: [
-        circle,
-        "communityId",
-      ],
-    },
-    spaceId: {
-      propDefinition: [
-        circle,
-        "spaceId",
-        ({ communityId }) => ({
-          communityId,
-        }),
-      ],
-      optional: true,
-    },
+    ...common.props,
     postId: {
       propDefinition: [
-        circle,
+        common.props.circle,
         "postId",
         ({
           communityId, spaceId,
@@ -49,14 +26,10 @@ export default {
     },
   },
   methods: {
-    _getLastId() {
-      return this.db.get("lastId") || 0;
-    },
-    _setLastId(lastId) {
-      return this.db.set("lastId", lastId);
-    },
-    async startEvent(maxResults = false) {
-      const lastId = this._getLastId();
+    ...common.methods,
+    async getResponse({
+      maxResults, lastId,
+    }) {
       const response = await this.circle.listComments({
         params: {
           community_id: this.communityId,
@@ -71,15 +44,10 @@ export default {
         this._setLastId(response[0].id);
       }
 
-      const responseArray = response.filter((item) => item.id > lastId);
-
-      for (const event of responseArray.reverse()) {
-        this.$emit(event, {
-          id: event.id,
-          summary: `New comment with ID: ${event.id}`,
-          ts: Date.parse(event.created_at),
-        });
-      }
+      return response.filter((item) => item.id > lastId);
+    },
+    getSummary({ id }) {
+      return `New comment with ID: ${id}`;
     },
   },
   hooks: {
